or1k: allow get/set fcscr in userspace
FPU control status register setting and getting should be allowed in
use space as is allowed in all other architectures.  However, on
openrisc it is only alowed in privilaged mode.

Currently its not allowed but we need it for glibc.
Needs a spec change of patches to the kernel but this
gets things working for testing.